Add state for About fragment, ViewModel and refactor
